requirements.txt
astroid==2.3.3
click==7.1.1
colorama==0.4.3
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.3
Pillow==7.1.1
pylint==2.4.4
six==1.14.0
torch==1.4.0+cpu
torchvision==0.5.0+cpu
Werkzeug==1.0.1
wrapt==1.11.2
